backend/pyproject.toml
[project]
name = "baja-inspection-backend"
version = "0.1.0"
description = "FastAPI backend serving YOLO paint-defect inference for the Baja chassis inspection chamber"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "pydantic>=2.6",
    "pillow>=10.2",
    "numpy>=1.26",
    "ultralytics>=8.1",
    "supabase>=2.4",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "huggingface_hub>=0.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
